termionix_telnetcodec

Struct TelnetCodec

pub struct TelnetCodec { /* private fields */ }

A codec for handling the Telnet sidechannel, providing functionality to encode and decode Telnet messages.

TelnetCodec is responsible for managing the state and buffers required for processing data when implementing the Telnet sidechannel. It is typically used in conjunction with asynchronous server libraries to handle the transmission and reception of Telnet messages over a connection.

This struct is typically paired with a transport-like implementation to facilitate stream I/O management for the Telnet sidechannel.

Implementations

impl TelnetCodec

pub fn new() -> TelnetCodec

Creates a new instance of TelnetCodec.

This constructor initializes the TelnetCodec struct with default settings:

encoder_buffer: A buffer for encoding data, initialized to its default state.

decoder_buffer: A buffer for decoding data, initialized to its default state.

decoder_state: The initial state of the decoder, set to DecoderState::NormalData.

options: A set of Telnet options, initialized with default values.

Returns

A TelnetCodec instance ready for use.

Example

use termionix_telnetcodec::TelnetCodec;

let codec = TelnetCodec::new();

pub fn has_pending_responses(&self) -> bool

Check if there are pending sidechannel responses

pub fn flush_responses( &mut self, dst: &mut BytesMut, ) -> Result<(), TelnetCodecError>

Encode all pending responses into buffer

pub fn is_supported_local(&self, option: TelnetOption) -> bool

Checks if we support the given option locally

pub fn is_supported_remote(&self, option: TelnetOption) -> bool

Checks if we support the given option remotely

pub fn is_enabled_local(&self, option: TelnetOption) -> bool

Checks if a specific Telnet option is enabled locally.

Parameters

option: The TelnetOption to check the local enablement status for.

Returns

bool: Returns true if the specified Telnet option is enabled locally, otherwise false.

Example

use termionix_telnetcodec::{TelnetCodec, TelnetOption};

let codec = TelnetCodec::new();
let is_enabled = codec.is_enabled_local(TelnetOption::Echo);
if is_enabled {
    println!("The Echo option is enabled locally.");
} else {
    println!("The Echo option is not enabled locally.");
}

Note

This function uses the local_enabled method of the options field to determine the enablement status.

pub fn is_enabled_remote(&self, option: TelnetOption) -> bool

Checks if a specific Telnet option is enabled on the remote side.

Parameters

option: The TelnetOption to check for its enabled status.

Returns

bool: Returns true if the given Telnet option is enabled on the remote side, otherwise returns false.

Example

use termionix_telnetcodec::{TelnetCodec, TelnetOption};

let codec = TelnetCodec::new();
let is_enabled = codec.is_enabled_remote(TelnetOption::Echo);
if is_enabled {
    println!("The Echo option is enabled locally.");
} else {
    println!("The Echo option is not enabled locally.");
}

pub fn enable_local(&mut self, option: TelnetOption) -> Option<TelnetFrame>

Request to enable a Telnet option locally (we will send WILL).

This initiates the negotiation process using the Q-method state machine. If negotiation is needed, returns a frame that should be sent to the remote side.

Parameters

option: The TelnetOption to enable locally.

Returns

Some(TelnetFrame): A negotiation frame to send to the remote side.

None: No negotiation needed (option already enabled or not supported).

Example

use termionix_telnetcodec::{TelnetCodec, TelnetOption};

let mut codec = TelnetCodec::new();
if let Some(frame) = codec.enable_local(TelnetOption::Echo) {
    // Send frame to remote side
}

pub fn disable_local(&mut self, option: TelnetOption) -> Option<TelnetFrame>

Request to disable a Telnet option locally (we will send WONT).

This initiates the negotiation process using the Q-method state machine. If negotiation is needed, returns a frame that should be sent to the remote side.

Parameters

option: The TelnetOption to disable locally.

Returns

Some(TelnetFrame): A negotiation frame to send to the remote side.

None: No negotiation needed (option already disabled).

Example

use termionix_telnetcodec::{TelnetCodec, TelnetOption};

let mut codec = TelnetCodec::new();
if let Some(frame) = codec.disable_local(TelnetOption::Echo) {
    // Send frame to remote side
}

pub fn enable_remote(&mut self, option: TelnetOption) -> Option<TelnetFrame>

Request to enable a Telnet option on the remote side (we will send DO).

This initiates the negotiation process using the Q-method state machine. If negotiation is needed, returns a frame that should be sent to the remote side.

Parameters

option: The TelnetOption to enable on the remote side.

Returns

Some(TelnetFrame): A negotiation frame to send to the remote side.

None: No negotiation needed (option already enabled or not supported).

Example

use termionix_telnetcodec::{TelnetCodec, TelnetOption};

let mut codec = TelnetCodec::new();
if let Some(frame) = codec.enable_remote(TelnetOption::SuppressGoAhead) {
    // Send frame to remote side
}

pub fn disable_remote(&mut self, option: TelnetOption) -> Option<TelnetFrame>

Request to disable a Telnet option on the remote side (we will send DONT).

This initiates the negotiation process using the Q-method state machine. If negotiation is needed, returns a frame that should be sent to the remote side.

Parameters

option: The TelnetOption to disable on the remote side.

Returns

Some(TelnetFrame): A negotiation frame to send to the remote side.

None: No negotiation needed (option already disabled).

Example

use termionix_telnetcodec::{TelnetCodec, TelnetOption};

let mut codec = TelnetCodec::new();
if let Some(frame) = codec.disable_remote(TelnetOption::Echo) {
    // Send frame to remote side
}